Removed GNU -Wno-sign-compare

+++ src/core/mp_gf2n.hpp
@@ -805,7 +805,7 @@ public:
             return;
         }
 
-        size_t in_used = in1.size();
+        int in_used = in1.size();
         if (0 == in_used || (bits >= (in_used * std::numeric_limits<T>::digits))) {
             out.resize(0);
             return;
