+++ src/crypto/aes_fpe_ff1.hpp
@@ -105,10 +105,10 @@ public:
         int Slen = 16 + cnt * 16;
         phantom_vector<uint8_t> S(Slen);
         phantom_vector<uint8_t> vecBytes;
-        for (int i = 0; i < ff1_rounds; i++) {
+        for (size_t i = 0; i < ff1_rounds; i++) {
 
             // v
-            int m = (i & 1)? v: u;
+            size_t m = (i & 1) ? v : u;
 
             // i
             Q[tweaklen + pad] = i & 0xff;
@@ -296,7 +296,7 @@ private:
 
         std::copy(R.begin(), R.begin() + 16, S.begin());
 
-        for (uint32_t j=1; j <= cnt; j++) {
+        for (int j=1; j <= cnt; j++) {
             std::copy(R.begin(), R.begin() + 16, tmp.begin());
 
 #if PLATFORM_BYTE_ORDER == IS_LITTLE_ENDIAN
